LKVProcessGraphic and IoT Central adapter (#10)
These changes include

addition of an LKVProcessGraphic Consume card

an IoT Central adapter to support its consumption in IoT Central companion apps

a TSI adapter to integrate with the LineChartCard.

As part of this, I have also refined a few interfaces (IBaseAdapter, IAuthService, ConsumeCardProps), and moved a lot of code around to create a more compact folder structure under src. There is a lot of code here, and some of it will need to be changed as we fill out common components (e.g. title as an optional prop on BaseCard). I believe these additions to be mostly beneficial, but not final implementations.